@@ -7,6 +7,7 @@ using Autofac.Extensions.DependencyInjection;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace ApiGateWay
@@ -15,18 +16,19 @@ namespace ApiGateWay
     {
         public static void Main(string[] args)
         {
-            CreateWebHostBuilder(args).Build().Run();
+            CreateHostBuilder(args).Build().Run();
         }
-
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration((hostContext, config) =>
-                {
-                    config.SetBasePath(Directory.GetCurrentDirectory());
-                    //获取oxygen配置节
-                    config.AddJsonFile("oxygen.json");
-                })
-                .UseStartup<Startup>()
-                .ConfigureServices(services => services.AddAutofac());
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+           Host.CreateDefaultBuilder(args)
+               .ConfigureWebHostDefaults(webBuilder =>
+               {
+                   webBuilder.ConfigureAppConfiguration((hostContext, config) =>
+                   {
+                       config.SetBasePath(Directory.GetCurrentDirectory());
+                       //获取oxygen配置节
+                       config.AddJsonFile("oxygen.json");
+                   });
+                   webBuilder.UseStartup<Startup>();
+               }).UseServiceProviderFactory(new AutofacServiceProviderFactory());
     }
 }
